musl support (#787)
- Test also on Alpine, which has muslc rather than glibc
- Added the required (limited)  changes in the source code.

+++ src/storm/modelchecker/lexicographic/lexicographicModelCheckerHelper.cpp
@@ -19,11 +19,11 @@ namespace helper {
 namespace lexicographic {
 
 template<typename SparseModelType, typename ValueType, bool Nondeterministic>
-std::pair<std::shared_ptr<storm::transformer::DAProduct<SparseModelType>>, std::vector<uint>>
+std::pair<std::shared_ptr<storm::transformer::DAProduct<SparseModelType>>, std::vector<uint64_t>>
 lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::getCompleteProductModel(const SparseModelType& model,
                                                                                                        CheckFormulaCallback const& formulaChecker) {
     storm::logic::ExtractMaximalStateFormulasVisitor::ApToFormulaMap extracted;
-    std::vector<uint> acceptanceConditions;
+    std::vector<uint64_t> acceptanceConditions;
 
     // Get the big product automton for all subformulae
     std::shared_ptr<storm::automata::DeterministicAutomaton> productAutomaton = spothelper::ltl2daSpotProduct(this->formula, extracted, acceptanceConditions);
@@ -60,7 +60,7 @@ lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::g
 template<typename SparseModelType, typename ValueType, bool Nondeterministic>
 std::pair<storm::storage::MaximalEndComponentDecomposition<ValueType>, std::vector<std::vector<bool>>>
 lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::getLexArrays(
-    std::shared_ptr<storm::transformer::DAProduct<productModelType>> productModel, std::vector<uint>& acceptanceConditions) {
+    std::shared_ptr<storm::transformer::DAProduct<productModelType>> productModel, std::vector<uint64_t>& acceptanceConditions) {
     storm::storage::BitVector allowed(productModel->getProductModel().getTransitionMatrix().getRowGroupCount(), true);
     // get MEC decomposition
     storm::storage::MaximalEndComponentDecomposition<ValueType> mecs(productModel->getProductModel().getTransitionMatrix(),
@@ -79,7 +79,7 @@ lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::g
     for (storm::storage::MaximalEndComponent& mec : mecs) {
         std::vector<storm::automata::AcceptanceCondition::acceptance_expr::ptr> sprime;
         std::vector<bool> bsccAccepting;
-        for (uint i = 0; i < acceptanceConditions.size() - 1; i++) {
+        for (uint64_t i = 0; i < acceptanceConditions.size() - 1; i++) {
             // copy the current list of Streett-pairs that can be fulfilled together
             std::vector<storm::automata::AcceptanceCondition::acceptance_expr::ptr> sprimeTemp(sprime);
             // add the new pairs (for the new condition) that should be checked now
@@ -111,7 +111,7 @@ MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelpe
     // Add first new states for each MEC
     auto stStateResult = addSinkStates(mecs, productModel);
     storm::storage::SparseMatrix<ValueType> newMatrixWithNewStates = stStateResult.first;
-    std::map<uint, uint_fast64_t> bccToStStateMapping = stStateResult.second;
+    std::map<uint64_t, uint64_t> bccToStStateMapping = stStateResult.second;
 
     // eliminate the MECs (collapse them into one state)
     storm::transformer::EndComponentEliminator<ValueType> eliminator = storm::transformer::EndComponentEliminator<ValueType>();
@@ -123,12 +123,12 @@ MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelpe
     std::vector<std::vector<bool>> bccLexArrayCurrent(mecLexArray);
     // prepare the result (one reachability probability for each objective)
     MDPSparseModelCheckingHelperReturnType<ValueType> retResult(std::vector<ValueType>(mecLexArray[0].size()));
-    std::vector<uint_fast64_t> compressedToReducedMapping(compressionResult.matrix.getColumnCount());
+    std::vector<uint64_t> compressedToReducedMapping(compressionResult.matrix.getColumnCount());
     std::iota(std::begin(compressedToReducedMapping), std::end(compressedToReducedMapping), 0);
     storm::storage::SparseMatrix<ValueType> transitionMatrix = compressionResult.matrix;
 
     // check reachability for each condition and restrict the model to optimal choices
-    for (uint condition = 0; condition < mecLexArray[0].size(); condition++) {
+    for (uint64_t condition = 0; condition < mecLexArray[0].size(); condition++) {
         // get the goal-states for this objective (i.e. the st-states of the MECs where the objective can be fulfilled
         storm::storage::BitVector psiStates = getGoodStates(mecs, bccLexArrayCurrent, compressionResult.oldToNewStateMapping, condition,
                                                             transitionMatrix.getColumnCount(), compressedToReducedMapping, bccToStStateMapping);
@@ -138,7 +138,7 @@ MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelpe
         }
 
         // solve the reachability query for this set of goal states
-        std::vector<uint_fast64_t> newInitalStates;
+        std::vector<uint64_t> newInitalStates;
         auto res =
             solveOneReachability(newInitalStates, psiStates, transitionMatrix, originalMdp, compressedToReducedMapping, compressionResult.oldToNewStateMapping);
         if (newInitalStates.empty()) {
@@ -149,10 +149,10 @@ MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelpe
 
         // create a reduced subsystem that only contains the optimal actions for this objective
         auto subsystem = getReducedSubsystem(transitionMatrix, res, newInitalStates, psiStates);
-        std::vector<uint_fast64_t> compressedToReducedMappingTemp;
+        std::vector<uint64_t> compressedToReducedMappingTemp;
         compressedToReducedMappingTemp.reserve(compressedToReducedMapping.size());
         // create a new mapping for the states
-        for (uint_fast64_t newState : subsystem.newToOldStateIndexMapping) {
+        for (uint64_t newState : subsystem.newToOldStateIndexMapping) {
             compressedToReducedMappingTemp.push_back(compressedToReducedMapping[newState]);
         }
         compressedToReducedMapping = compressedToReducedMappingTemp;
@@ -255,20 +255,20 @@ bool lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministi
 template<typename SparseModelType, typename ValueType, bool Nondeterministic>
 storm::storage::BitVector lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::getGoodStates(
     storm::storage::MaximalEndComponentDecomposition<ValueType> const& bcc, std::vector<std::vector<bool>> const& bccLexArray,
-    std::vector<uint_fast64_t> const& oldToNewStateMapping, uint const& condition, uint const numStates,
-    std::vector<uint_fast64_t> const& compressedToReducedMapping, std::map<uint, uint_fast64_t> const& bccToStStateMapping) {
+    std::vector<uint64_t> const& oldToNewStateMapping, uint64_t const& condition, uint64_t const numStates,
+    std::vector<uint64_t> const& compressedToReducedMapping, std::map<uint64_t, uint64_t> const& bccToStStateMapping) {
     STORM_LOG_ASSERT(!bccLexArray.empty(), "Lex-Array is empty!");
     STORM_LOG_ASSERT(condition < bccLexArray[0].size(), "Condition is not in Lex-Array!");
-    std::vector<uint_fast64_t> goodStates;
-    for (uint i = 0; i < bcc.size(); i++) {
+    std::vector<uint64_t> goodStates;
+    for (uint64_t i = 0; i < bcc.size(); i++) {
         std::vector<bool> const& bccLex = bccLexArray[i];
         if (bccLex[condition]) {
-            uint_fast64_t bccStateOld = bccToStStateMapping.at(i);
-            uint_fast64_t bccState = oldToNewStateMapping[bccStateOld];
+            uint64_t bccStateOld = bccToStStateMapping.at(i);
+            uint64_t bccState = oldToNewStateMapping[bccStateOld];
             auto pointer = std::find(compressedToReducedMapping.begin(), compressedToReducedMapping.end(), bccState);
             if (pointer != compressedToReducedMapping.end()) {
                 // We have to check whether the state has already been removed
-                uint_fast64_t bccStateReduced = pointer - compressedToReducedMapping.begin();
+                uint64_t bccStateReduced = pointer - compressedToReducedMapping.begin();
                 goodStates.push_back(bccStateReduced);
             }
         }
@@ -278,8 +278,8 @@ storm::storage::BitVector lexicographicModelCheckerHelper<SparseModelType, Value
 
 template<typename SparseModelType, typename ValueType, bool Nondeterministic>
 MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::solveOneReachability(
-    std::vector<uint_fast64_t>& newInitalStates, storm::storage::BitVector const& psiStates, storm::storage::SparseMatrix<ValueType> const& transitionMatrix,
-    SparseModelType const& originalMdp, std::vector<uint_fast64_t> const& compressedToReducedMapping, std::vector<uint_fast64_t> const& oldToNewStateMapping) {
+    std::vector<uint64_t>& newInitalStates, storm::storage::BitVector const& psiStates, storm::storage::SparseMatrix<ValueType> const& transitionMatrix,
+    SparseModelType const& originalMdp, std::vector<uint64_t> const& compressedToReducedMapping, std::vector<uint64_t> const& oldToNewStateMapping) {
     Environment env;
     // A reachability condition "F x" is transformed to "true U x"
     // phi states are all states
@@ -289,7 +289,7 @@ MDPSparseModelCheckingHelperReturnType<ValueType> lexicographicModelCheckerHelpe
     // Get initial states in the compressed model
     storm::storage::BitVector const& originalInitialStates = originalMdp.getInitialStates();
 
-    uint pos = 0;
+    uint64_t pos = 0;
     while (originalInitialStates.getNextSetIndex(pos) != originalInitialStates.size()) {
         pos = originalInitialStates.getNextSetIndex(pos) + 1;
         auto pointer = std::find(compressedToReducedMapping.begin(), compressedToReducedMapping.end(), oldToNewStateMapping[pos - 1]);
@@ -310,23 +310,23 @@ template<typename SparseModelType, typename ValueType, bool Nondeterministic>
 typename lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::SubsystemReturnType
 lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::getReducedSubsystem(
     storm::storage::SparseMatrix<ValueType> const& transitionMatrix, MDPSparseModelCheckingHelperReturnType<ValueType> const& reachabilityResult,
-    std::vector<uint_fast64_t> const& newInitalStates, storm::storage::BitVector const& goodStates) {
-    std::vector<std::vector<uint>> optimalActions;
-    std::vector<uint_fast64_t> keptActions;
-    std::vector<uint_fast64_t> const& rowGroupIndices = transitionMatrix.getRowGroupIndices();
+    std::vector<uint64_t> const& newInitalStates, storm::storage::BitVector const& goodStates) {
+    std::vector<std::vector<uint64_t>> optimalActions;
+    std::vector<uint64_t> keptActions;
+    std::vector<uint64_t> const& rowGroupIndices = transitionMatrix.getRowGroupIndices();
 
     // iterate over the states
-    for (uint currentState = 0; currentState < reachabilityResult.values.size(); currentState++) {
-        std::vector<uint> goodActionsForState;
-        uint_fast64_t bestAction = reachabilityResult.scheduler->getChoice(currentState).getDeterministicChoice();
+    for (uint64_t currentState = 0; currentState < reachabilityResult.values.size(); currentState++) {
+        std::vector<uint64_t> goodActionsForState;
+        uint64_t bestAction = reachabilityResult.scheduler->getChoice(currentState).getDeterministicChoice();
         // determine the value of the best action
         ValueType bestActionValue(0);
-        for (const storm::storage::MatrixEntry<uint_fast64_t, ValueType>& rowEntry : transitionMatrix.getRow(rowGroupIndices[currentState] + bestAction)) {
+        for (const storm::storage::MatrixEntry<uint64_t, ValueType>& rowEntry : transitionMatrix.getRow(rowGroupIndices[currentState] + bestAction)) {
             bestActionValue += rowEntry.getValue() * reachabilityResult.values[rowEntry.getColumn()];
         }
-        uint lastAction = rowGroupIndices[currentState] + transitionMatrix.getRowGroupSize(currentState);
+        uint64_t lastAction = rowGroupIndices[currentState] + transitionMatrix.getRowGroupSize(currentState);
         // iterate over all actions in this state to find those that are also optimal
-        for (uint action = rowGroupIndices[currentState]; action < lastAction; action++) {
+        for (uint64_t action = rowGroupIndices[currentState]; action < lastAction; action++) {
             ValueType actionValue(0);
             for (const auto& rowEntry : transitionMatrix.getRow(action)) {
                 actionValue += rowEntry.getValue() * reachabilityResult.values[rowEntry.getColumn()];
@@ -353,33 +353,33 @@ lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::g
 }
 
 template<typename SparseModelType, typename ValueType, bool Nondeterministic>
-std::pair<storm::storage::SparseMatrix<ValueType>, std::map<uint, uint_fast64_t>>
+std::pair<storm::storage::SparseMatrix<ValueType>, std::map<uint64_t, uint64_t>>
 lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::addSinkStates(
     storm::storage::MaximalEndComponentDecomposition<ValueType> const& mecs,
     std::shared_ptr<storm::transformer::DAProduct<SparseModelType>> const& productModel) {
     storm::storage::SparseMatrix<ValueType> const matrix = productModel->getProductModel().getTransitionMatrix();
-    uint countNewRows = 0;
-    std::map<uint_fast64_t, uint> stateToMEC;
-    for (uint i = 0; i < mecs.size(); i++) {
+    uint64_t countNewRows = 0;
+    std::map<uint64_t, uint64_t> stateToMEC;
+    for (uint64_t i = 0; i < mecs.size(); i++) {
         auto bcc = mecs[i];
         countNewRows += bcc.size() + 1;
         for (auto const& stateAndChoices : bcc) {
             // create an assignment of states to MECS
-            uint_fast64_t bccState = stateAndChoices.first;
+            uint64_t bccState = stateAndChoices.first;
             stateToMEC[bccState] = i;
         }
     }
     // create a new transition matrix
     storm::storage::SparseMatrixBuilder<ValueType> builder(matrix.getRowCount() + countNewRows, matrix.getColumnCount() + mecs.size(),
                                                            matrix.getEntryCount() + countNewRows, false, true, matrix.getRowGroupCount());
-    uint_fast64_t numRowGroups = matrix.getColumnCount() + mecs.size();
-    std::map<uint, uint_fast64_t> sTstatesForBCC;
-    uint_fast64_t newestRowGroup = matrix.getRowGroupCount();
-    uint_fast64_t newRow = 0;
-    uint_fast64_t oldRowCounting = 0;
+    uint64_t numRowGroups = matrix.getColumnCount() + mecs.size();
+    std::map<uint64_t, uint64_t> sTstatesForBCC;
+    uint64_t newestRowGroup = matrix.getRowGroupCount();
+    uint64_t newRow = 0;
+    uint64_t oldRowCounting = 0;
     auto oldRowGroupIndices = matrix.getRowGroupIndices();
     // iterate over the row groups (aka states) of the old transition matrix
-    for (uint_fast64_t newRowGroup = 0; newRowGroup < matrix.getColumnCount(); ++newRowGroup) {
+    for (uint64_t newRowGroup = 0; newRowGroup < matrix.getColumnCount(); ++newRowGroup) {
         // create a new row group (aka state) for this
         builder.newRowGroup(newRow);
         // iterate over the rows of this row group (aka transitions from this state)
@@ -402,7 +402,7 @@ lexicographicModelCheckerHelper<SparseModelType, ValueType, Nondeterministic>::a
     }
     // add new row groups (aka states) for the new sink states to the transition matrix
     // only possible action of those is a self-loop
-    for (uint_fast64_t newRowGroup = matrix.getColumnCount(); newRowGroup < numRowGroups; newRowGroup++) {
+    for (uint64_t newRowGroup = matrix.getColumnCount(); newRowGroup < numRowGroups; newRowGroup++) {
         builder.newRowGroup(newRow);
         builder.addNextValue(newRow, newRowGroup, storm::utility::one<ValueType>());
         newRow++;
